+++ CHANGELOG.md
@@ -1,5 +1,35 @@
 
 
+## [0.1.1-alpha.5](https://github.com/MinhOmega/npm-increaser-downloads/compare/v0.1.1-alpha.4...v0.1.1-alpha.5) (2024-09-14)
+
+## [0.1.1-alpha.4](https://github.com/MinhOmega/npm-increaser-downloads/compare/v0.1.1-alpha.3...v0.1.1-alpha.4) (2024-09-14)
+
+
+### Bug Fixes
+
+* crash on running command line ([517edd9](https://github.com/MinhOmega/npm-increaser-downloads/commit/517edd9128aeb94e74055ace983549e81283a279))
+
+## [0.1.1-alpha.3](https://github.com/MinhOmega/npm-increaser-downloads/compare/v0.1.1-alpha.2...v0.1.1-alpha.3) (2024-09-14)
+
+
+### Features
+
+* implement command-line arguments ([b41ff50](https://github.com/MinhOmega/npm-increaser-downloads/commit/b41ff5066eed12ab80a507c049d069d61f69924e))
+
+## [0.1.1-alpha.2](https://github.com/MinhOmega/npm-increaser-downloads/compare/v0.1.1-alpha.1...v0.1.1-alpha.2) (2024-09-14)
+
+
+### Bug Fixes
+
+* crash on running command ([5896361](https://github.com/MinhOmega/npm-increaser-downloads/commit/58963615617ee19cb33b9812ade42f42f2ef3f57))
+
+## [0.1.1-alpha.1](https://github.com/MinhOmega/npm-increaser-downloads/compare/v0.1.0...v0.1.1-alpha.1) (2024-09-14)
+
+
+### Features
+
+* implement new package prompt and new cli ([28d9641](https://github.com/MinhOmega/npm-increaser-downloads/commit/28d9641d830836c8fe0c7c26f3e4d16aa0ff6419))
+
 ## [0.1.0](https://github.com/MinhOmega/npm-increaser-downloads/compare/v0.0.2...v0.1.0) (2024-09-13)